Separated unit test and made tests a package
