+++ tests/__init__.py
@@ -1,0 +1,4 @@
+from pathlib import Path
+
+BASE_DIR = Path(__file__).parent.parent
+ASSETS_DIR = BASE_DIR / "assets"
